[fix]calcコード全体修正
