+++ src/batch/calc/normalize.py
@@ -1,0 +1,62 @@
+import re
+import unicodedata
+
+
+class Normalize:
+    def __init__(self, sentence):
+        self.sentence = sentence
+
+    def normalize(self):
+        return self.normalize_neologd()
+
+    def unicode_normalize(self, cls, s):
+        pt = re.compile('([{}]+)'.format(cls))
+
+        def norm(c):
+            return unicodedata.normalize('NFKC', c) if pt.match(c) else c
+
+        s = ''.join(norm(x) for x in re.split(pt, s))
+        s = re.sub('－', '-', s)
+        return s
+
+    def remove_extra_spaces(self, s):
+        s = re.sub('[ 　]+', ' ', s)
+        blocks = ''.join(('\u4E00-\u9FFF',  # CJK UNIFIED IDEOGRAPHS
+                          '\u3040-\u309F',  # HIRAGANA
+                          '\u30A0-\u30FF',  # KATAKANA
+                          '\u3000-\u303F',  # CJK SYMBOLS AND PUNCTUATION
+                          '\uFF00-\uFFEF'   # HALFWIDTH AND FULLWIDTH FORMS
+                          ))
+        basic_latin = '\u0000-\u007F'
+
+        def remove_space_between(cls1, cls2, s):
+            p = re.compile('([{}]) ([{}])'.format(cls1, cls2))
+            while p.search(s):
+                s = p.sub(r'\1\2', s)
+            return s
+
+        s = remove_space_between(blocks, blocks, s)
+        s = remove_space_between(blocks, basic_latin, s)
+        s = remove_space_between(basic_latin, blocks, s)
+        return s
+
+    def normalize_neologd(self):
+        s = self.sentence
+        s = s.strip()
+        s = self.unicode_normalize('０-９Ａ-Ｚａ-ｚ｡-ﾟ', s)
+
+        def maketrans(f, t):
+            return {ord(x): ord(y) for x, y in zip(f, t)}
+
+        s = re.sub('[˗֊‐‑‒–⁃⁻₋−]+', '-', s)  # normalize hyphens
+        s = re.sub('[﹣－ｰ—―─━ー]+', 'ー', s)  # normalize choonpus
+        s = re.sub('[~∼∾〜〰～]+', '〜', s)  # normalize tildes (modified by Isao Sonobe)
+        s = s.translate(
+            maketrans('!"#$%&\'()*+,-./:;<=>?@[¥]^_`{|}~｡､･｢｣',
+                      '！”＃＄％＆’（）＊＋，－．／：；＜＝＞？＠［￥］＾＿｀｛｜｝〜。、・「」'))
+        s = self.remove_extra_spaces(s)
+        s = self.unicode_normalize('！”＃＄％＆’（）＊＋，－．／：；＜＞？＠［￥］＾＿｀｛｜｝〜', s)  # keep ＝,・,「,」
+        s = re.sub('[’]', '\'', s)
+        s = re.sub('[”]', '"', s)
+        s = s.upper()
+        return s
